Refactor project structure
Adding directories, making only required items publicly available.
