+++ src/database/db.rs
@@ -1,9 +1,10 @@
-use crate::state::State;
 use anyhow::Result;
 use std::{
     fs::OpenOptions,
     io::{BufReader, BufWriter},
 };
+
+use crate::state::states::State;
 
 const DB_PATH: &str = "tasks.json";
