+++ src/database/mod.rs
@@ -1,0 +1,3 @@
+pub mod db;
+
+pub use db::{load, save};
