+++ src/event_handler/event_handlers.rs
@@ -1,5 +1,6 @@
-use crate::app::{App, InputMode};
 use crossterm::event::{KeyCode, KeyEvent, KeyEventKind};
+
+use crate::application::{App, InputMode};
 
 pub struct EventHandler;
 impl EventHandler {
