+++ src/lib.rs
@@ -1,7 +1,5 @@
-pub mod app;
+pub mod application;
 pub mod database;
 pub mod event_handler;
 pub mod state;
-pub mod tasks;
-pub mod terminal;
-pub mod ui;
+pub mod user_interface;
